+++ payment-service/payment/views.py
@@ -1,41 +1,54 @@
-from django.shortcuts import render
-
-# Create your views here.
-# payment/views.py
 import traceback
+
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
+
+from shared.authentication import JWTAuthentication
+from shared.permissions import IsCustomer
 
 from .serializers import (
     CreatePaymentSerializer,
     UpdatePaymentStatusSerializer,
 )
-
 from .services import PaymentService
 
-
 class PaymentListCreateView(APIView):
-    """
-    GET  /api/v1/payments/
-    POST /api/v1/payments/
-    """
+
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsCustomer]
 
     def get(self, request):
-        payments = PaymentService.get_all_payments()
-        return Response(payments, status=status.HTTP_200_OK)
+        try:
 
+            payments = PaymentService.get_payments_by_user(
+                request.user["user_id"]
+            )
+
+            return Response(
+                payments,
+                status=status.HTTP_200_OK
+            )
+
+        except Exception:
+            return Response(
+                {"error": "Internal server error"},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR
+            )
+        
     def post(self, request):
+
         serializer = CreatePaymentSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
         try:
+            auth_header = request.headers.get("Authorization")
+            token = auth_header.split()[1]
             payment = PaymentService.create_payment(
                 order_id=serializer.validated_data["order_id"],
-                user_id=serializer.validated_data["user_id"],
-                amount=serializer.validated_data["amount"],
-                currency=serializer.validated_data["currency"],
+                user_id=request.user["user_id"],
                 payment_method=serializer.validated_data["payment_method"],
+                token = token
             )
 
             return Response(
@@ -44,6 +57,7 @@ class PaymentListCreateView(APIView):
             )
 
         except ValueError as e:
+
             return Response(
                 {"error": str(e)},
                 status=status.HTTP_400_BAD_REQUEST,
@@ -51,6 +65,7 @@ class PaymentListCreateView(APIView):
 
         except Exception as e:
             traceback.print_exc()
+
             return Response(
                 {"error": str(e)},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
@@ -58,10 +73,9 @@ class PaymentListCreateView(APIView):
 
 
 class PaymentDetailView(APIView):
-    """
-    GET /api/v1/payments/<payment_id>/
-    DELETE /api/v1/payments/<payment_id>/
-    """
+
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsCustomer]
 
     def get(self, request, payment_id):
 
@@ -73,11 +87,32 @@ class PaymentDetailView(APIView):
                 status=status.HTTP_404_NOT_FOUND,
             )
 
+        if payment["user_id"] != request.user["user_id"]:
+            return Response(
+                {"error": "Permission denied"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         return Response(payment)
 
     def delete(self, request, payment_id):
 
+        payment = PaymentService.get_payment(payment_id)
+
+        if not payment:
+            return Response(
+                {"error": "Payment not found"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
+        if payment["user_id"] != request.user["user_id"]:
+            return Response(
+                {"error": "Permission denied"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         try:
+
             PaymentService.delete_payment(payment_id)
 
             return Response(
@@ -85,30 +120,49 @@ class PaymentDetailView(APIView):
             )
 
         except ValueError as e:
+
             return Response(
                 {"error": str(e)},
                 status=status.HTTP_404_NOT_FOUND,
             )
 
+        except Exception:
+
+            return Response(
+                {"error": "Internal server error"},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            )
+
 
 class PaymentStatusView(APIView):
-    """
-    PUT /api/v1/payments/<payment_id>/status/
-    """
 
-    def put(self, request, payment_id):
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsCustomer]
+
+    def patch(self, request, payment_id):
 
         serializer = UpdatePaymentStatusSerializer(
             data=request.data
         )
 
-        if not serializer.is_valid(raise_exception=True):
+        serializer.is_valid(raise_exception=True)
+
+        payment = PaymentService.get_payment(payment_id)
+
+        if not payment:
             return Response(
-                {"error": serializer.errors},
-                status=status.HTTP_400_BAD_REQUEST,
+                {"error": "Payment not found"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
+        if payment["user_id"] != request.user["user_id"]:
+            return Response(
+                {"error": "Permission denied"},
+                status=status.HTTP_403_FORBIDDEN,
             )
 
         try:
+
             payment = PaymentService.update_status(
                 payment_id=payment_id,
                 status=serializer.validated_data["status"],
@@ -124,11 +178,14 @@ class PaymentStatusView(APIView):
             )
 
         except ValueError as e:
+
             return Response(
                 {"error": str(e)},
                 status=status.HTTP_404_NOT_FOUND,
             )
+
         except Exception:
+
             return Response(
                 {"error": "Internal server error"},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
@@ -136,9 +193,9 @@ class PaymentStatusView(APIView):
 
 
 class PaymentByOrderView(APIView):
-    """
-    GET /api/v1/payments/order/<order_id>/
-    """
+
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsCustomer]
 
     def get(self, request, order_id):
 
@@ -150,6 +207,12 @@ class PaymentByOrderView(APIView):
                 status=status.HTTP_404_NOT_FOUND,
             )
 
+        if payment["user_id"] != request.user["user_id"]:
+            return Response(
+                {"error": "Permission denied"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         return Response(
             payment,
             status=status.HTTP_200_OK,
@@ -157,12 +220,7 @@ class PaymentByOrderView(APIView):
 
 
 class HealthView(APIView):
-    """
-    GET /api/v1/payments/health/
-    """
-
     def get(self, request):
-
         return Response(
             PaymentService.health(),
             status=status.HTTP_200_OK,
